Disable pty-proxy on Windows
